upgrade cargo and deps
