+++ programs/xnft/Cargo.toml
@@ -19,5 +19,5 @@ default = []
 [dependencies]
 anchor-lang = { git = "https://github.com/coral-xyz/anchor" }
 anchor-spl = { git = "https://github.com/coral-xyz/anchor", features = ["metadata"] }
-mpl-token-metadata = { version = "1.3.4", features = ["no-entrypoint"] }
+mpl-token-metadata = { version = "1.4.3", features = ["no-entrypoint"] }
 solana-security-txt = "1.0.1"
