+++ programs/xnft/src/state/mod.rs
@@ -13,8 +13,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use anchor_lang::prelude::*;
-
 mod access;
 mod install;
 mod review;
@@ -24,48 +22,3 @@ pub use self::xnft::*; // use `self::` prefix to remove crate vs module ambiguit
 pub use access::*;
 pub use install::*;
 pub use review::*;
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, PartialEq, Eq)]
-pub enum Kind {
-    App,
-    Collection { pubkey: Pubkey },
-    Nft { pubkey: Pubkey },
-}
-
-impl Kind {
-    pub fn as_pubkey(&self) -> Pubkey {
-        match self {
-            Kind::App => Pubkey::default(),
-            Kind::Collection { pubkey } | Kind::Nft { pubkey } => *pubkey,
-        }
-    }
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug)]
-pub enum Tag {
-    None,
-    Defi,
-    Game,
-    Nfts,
-}
-
-#[cfg(test)]
-mod tests {
-    use anchor_lang::prelude::Pubkey;
-    use std::str::FromStr;
-
-    use super::Kind;
-
-    #[test]
-    fn kind_variant_as_bytes() {
-        let pk = Pubkey::from_str("BaHSGaf883GA3u8qSC5wNigcXyaScJLSBJZbALWvPcjs").unwrap();
-        let x = Kind::App.as_pubkey();
-        assert_eq!(x.as_ref(), Pubkey::default().as_ref());
-
-        let y = Kind::Collection { pubkey: pk }.as_pubkey();
-        assert_eq!(y.as_ref(), pk.as_ref());
-
-        let z = Kind::Nft { pubkey: pk }.as_pubkey();
-        assert_eq!(z.as_ref(), pk.as_ref());
-    }
-}
