+++ typescript/src/index.ts
@@ -21,6 +21,7 @@ export * from "./accounts";
 export * from "./addresses";
 export * from "./instructions";
 export * from "./types";
+export * from "./xnft";
 
 export const PROGRAM_ID = new PublicKey(
   "xnft5aaToUM4UFETUQfj7NUDUBdvYHTVhNFThEYTm55"
