+++ typescript/src/types.ts
@@ -18,10 +18,10 @@
 import type { IdlAccounts, IdlTypes } from "@project-serum/anchor";
 import type { Xnft } from "./xnft";
 
-export type Access = IdlAccounts<Xnft>["access"];
-export type Install = IdlAccounts<Xnft>["install"];
-export type Review = IdlAccounts<Xnft>["review"];
-export type xNFT = IdlAccounts<Xnft>["xnft"];
+export type AccessAccount = IdlAccounts<Xnft>["access"];
+export type InstallAccount = IdlAccounts<Xnft>["install"];
+export type ReviewAccount = IdlAccounts<Xnft>["review"];
+export type XnftAccount = IdlAccounts<Xnft>["xnft"];
 
 export type CreateXnftParameters = IdlTypes<Xnft>["CreateXnftParams"];
 export type UpdateXnftParameters = IdlTypes<Xnft>["UpdateParams"];
